requirements.txt
tensorflow
matplotlib
Pillow
beautifulsoup4 
requests
scipy